+++ src/common/components/FieldFiscalYear/index.js
@@ -1,0 +1,2 @@
+export { FieldFiscalYear } from './FieldFiscalYear';
+export { FieldFiscalYearContainer } from './FieldFiscalYearContainer';