+++ src/common/hooks/index.js
@@ -1,6 +1,7 @@
 export * from './useAddressCategories';
 export * from './useBatchGroup';
 export * from './useConfigsAdjustments';
+export * from './useFiscalYear';
 export * from './useFundDistributionValidation';
 export * from './useInvoice';
 export * from './useInvoiceMutation';
@@ -9,6 +10,7 @@ export * from './useInvoiceLineMutation';
 export * from './useOrderLine';
 export * from './useOrderLines';
 export * from './useOrders';
+export * from './usePayableFiscalYears';
 export * from './useVendors';
 export * from './useVoucherById';
 export * from './useVoucherByInvoiceId';